Make HAL defined interrupt handlers overrideable
