+++ esp-hal-common/ld/esp32s2/esp32s2.x
@@ -34,4 +34,5 @@ INCLUDE "stack.x"
 
 EXTERN(DefaultHandler);
 
+INCLUDE "weak_isrs.x"
 INCLUDE "device.x"
