+++ esp-hal-common/ld/esp32s3/esp32s3.x
@@ -48,4 +48,5 @@ INCLUDE "stack.x"
 
 EXTERN(DefaultHandler);
 
+INCLUDE "weak_isrs.x"
 INCLUDE "device.x"
