+++ esp-hal-common/ld/riscv/hal-defaults.x
@@ -1,3 +1,5 @@
+INCLUDE "weak_isrs.x"
+
 PROVIDE(DefaultHandler = EspDefaultHandler);
 
 PROVIDE(interrupt1 = DefaultHandler);
